+++ PseudocodeInterpreter/Source/ExtensionMethods.cs
