+++ PseudocodeInterpreter/Source/Objects/BooleanLiteral.cs
