+++ PseudocodeInterpreter/Source/Objects/Literal.cs
