+++ PseudocodeInterpreter/Source/Objects/StringLiteral.cs
